DL5YA_24-master/requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
h5py==3.10.0
kiwisolver==1.4.5
matplotlib
mpmath==1.3.0
numpy
h5py
packaging==23.2
Pillow==10.1.0
pip==23.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
setuptools==68.2.2
six==1.16.0
sympy==1.12
